add new hooks file with new hook, use new hook to add cache
